requirements.txt
absl-py==2.4.0
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.1.0
av==16.1.0
blinker==1.9.0
blis==1.3.3
brotli==1.2.0
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.5
cutlet==0.5.0
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
dashscope==1.25.12
datasets==3.4.1
decorator==5.2.1
demucs==4.0.1
diffusers==0.36.0
dill==0.3.8
docopt==0.6.2
dora_search==0.1.12
dyNET38==2.2
einops==0.8.2
einx==0.3.0
fastapi==0.129.2
ffmpeg==1.4
ffmpy==1.0.0
filelock==3.24.3
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2024.12.0
fugashi==1.5.2
gradio==6.6.0
gradio_client==2.1.0
greenlet==3.3.2
groovy==0.1.2
grpcio==1.78.0
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
jaconv==0.5.0
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.9
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
lightning==2.6.1
lightning-utilities==0.15.2
llvmlite==0.46.0
logger==1.4
loguru==0.7.3
lycoris_lora==3.4.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
matplotlib==3.10.1
mdurl==0.1.2
modelscope==1.34.0
mojimoji==0.0.13
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.16
murmurhash==1.0.15
nagisa==0.2.11
networkx==3.6.1
num2words==0.5.14
numba==0.64.0
numpy==2.4.2
omegaconf==2.3.0
onnxruntime==1.24.1
openunmix==1.3.0
orjson==3.11.7
packaging==26.0
pandas==3.0.1
peft==0.18.1
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py3langid==0.3.0
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
pypinyin==0.53.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.3
qwen-asr==0.0.6
qwen-omni-utils==0.0.9
regex==2026.2.19
requests==2.32.5
retrying==1.4.2
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
setuptools==82.0.0
shellingham==1.5.4
silero-vad==6.2.0
six==1.17.0
smart_open==7.5.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
soynlp==0.0.493
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
srt==3.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.52.1
submitit==1.5.4
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.13.3
torch==2.10.0+cu128
torchao==0.16.0
torchaudio==2.10.0
torchcodec==0.10.0
torchmetrics==1.8.2
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
treetable==0.2.6
triton-windows==3.3.1.post21
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unidic-lite==1.0.8
urllib3==2.6.3
uvicorn==0.41.0
vector-quantize-pytorch==1.27.21
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
win32_setctime==1.2.0
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
